backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi-sessions>=0.3.2",
    "fastapi[standard]>=0.115.12",
    "openai>=1.78.1",
    "pdfminer-six>=20250506",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
    "uuid>=1.30",
]
